+++ src/backend/termion.rs
@@ -84,15 +84,4 @@ impl EventHandler for Input {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn handle_tab() {
-        let evt = Event::Key(Key::Char('\t'));
-
-        let req = to_input_request(&evt);
-
-        assert!(req.is_none());
-    }
-}
+mod tests;
