+++ src/backend/termion/tests.rs
@@ -1,0 +1,10 @@
+use super::*;
+
+#[test]
+fn handle_tab() {
+    let evt = Event::Key(Key::Char('\t'));
+
+    let req = to_input_request(&evt);
+
+    assert!(req.is_none());
+}
